Added ICameraTarget, an interface for what the camera is allowed to point at. Currently, that is Graphics and Physics tokens.
Added the interface to the graphics and physics tokens.

Added logic to the client and server, and their base class game, to handle the camera target.

Camera Target can be set by calling SetTarget in the GameState class.

The player object now sets itself as a target if it is created on a client.